Move MockMetric to jdisc_core test package
