+++ jdisc_core/src/main/java/com/yahoo/jdisc/test/MockMetric.java
@@ -1,5 +1,5 @@
 // Copyright Verizon Media. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.docproc.jdisc.metric;
+package com.yahoo.jdisc.test;
 
 import com.yahoo.jdisc.Metric;
